+++ lib/src/side_sheet_component.dart
@@ -7,32 +7,41 @@ class SideSheet {
   /// ```
   static Future<dynamic> left(
       {
-
       ///Use this to pass any widget you want to display in the side sheet
       required Widget body,
       required BuildContext context,
 
       /// Use this to set the width of the side sheet
       double? width,
-      String barrierLabel: "Side Sheet",
+      String barrierLabel = "Side Sheet",
 
       /// Use barrierDismissible to dismiss the side sheet by tapping outside of it
-      bool barrierDismissible: true,
+      bool barrierDismissible = true,
 
       /// Use barrierColor to set the outside color of the side sheet when opened
       Color barrierColor = const Color(0xFF66000000),
 
+      /// Use sheetBorderRadius to set the outside border radius of the side sheet when opened
+      double sheetBorderRadius = 0,
+
+      /// Use sheetColor to set the color of the side sheet when opened
+      /// Default is white
+      Color sheetColor = Colors.white,
+
       /// Use transitionDuration to set the duration of the animation when the side sheet is opened or closed
       Duration transitionDuration = const Duration(milliseconds: 300)}) async {
     dynamic data = await _showSheetSide(
-        body: body,
-        width: width,
-        rightSide: false,
-        context: context,
-        barrierLabel: barrierLabel,
-        barrierDismissible: barrierDismissible,
-        barrierColor: barrierColor,
-        transitionDuration: transitionDuration);
+      body: body,
+      width: width,
+      rightSide: false,
+      context: context,
+      barrierLabel: barrierLabel,
+      barrierDismissible: barrierDismissible,
+      barrierColor: barrierColor,
+      sheetBorderRadius: sheetBorderRadius,
+      sheetColor: sheetColor,
+      transitionDuration: transitionDuration,
+    );
     if (data == null) return '';
 
     return data;
@@ -44,32 +53,41 @@ class SideSheet {
   /// ```
   static Future<dynamic> right(
       {
-
       ///Use this to pass any widget you want to display in the side sheet
       required Widget body,
       required BuildContext context,
 
       /// Use this to set the width of the side sheet
       double? width,
-      String barrierLabel: "Side Sheet",
+      String barrierLabel = "Side Sheet",
 
       /// Use barrierDismissible to dismiss the side sheet by tapping outside of it
-      bool barrierDismissible: true,
+      bool barrierDismissible = true,
 
       /// Use barrierColor to set the outside color of the side sheet when opened
       Color barrierColor = const Color(0xFF66000000),
 
+      /// Use sheetBorderRadius to set the outside border radius of the side sheet when opened
+      double sheetBorderRadius = 0,
+
+      /// Use sheetColor to set the color of the side sheet when opened
+      /// Default is white
+      Color sheetColor = Colors.white,
+
       /// Use transitionDuration to set the duration of the animation when the side sheet is opened or closed
       Duration transitionDuration = const Duration(milliseconds: 300)}) async {
     dynamic data = await _showSheetSide(
-        body: body,
-        width: width,
-        rightSide: true,
-        context: context,
-        barrierLabel: barrierLabel,
-        barrierDismissible: barrierDismissible,
-        barrierColor: barrierColor,
-        transitionDuration: transitionDuration);
+      body: body,
+      width: width,
+      rightSide: true,
+      context: context,
+      barrierLabel: barrierLabel,
+      barrierDismissible: barrierDismissible,
+      barrierColor: barrierColor,
+      sheetBorderRadius: sheetBorderRadius,
+      sheetColor: sheetColor,
+      transitionDuration: transitionDuration,
+    );
     if (data == null) return '';
 
     return data;
@@ -83,8 +101,18 @@ class SideSheet {
     required String barrierLabel,
     required bool barrierDismissible,
     required Color barrierColor,
+    required double sheetBorderRadius,
+    required Color sheetColor,
     required Duration transitionDuration,
   }) {
+    BorderRadius borderRadius = BorderRadius.only(
+      topLeft: rightSide ? Radius.circular(sheetBorderRadius) : Radius.zero,
+      bottomLeft: rightSide ? Radius.circular(sheetBorderRadius) : Radius.zero,
+      topRight: !rightSide ? Radius.circular(sheetBorderRadius) : Radius.zero,
+      bottomRight:
+          !rightSide ? Radius.circular(sheetBorderRadius) : Radius.zero,
+    );
+
     return showGeneralDialog(
       barrierLabel: barrierLabel,
       barrierDismissible: barrierDismissible,
@@ -97,8 +125,10 @@ class SideSheet {
           child: Material(
             elevation: 15,
             color: Colors.transparent,
+            borderRadius: borderRadius,
             child: Container(
-                color: Colors.white,
+                decoration: BoxDecoration(
+                    color: sheetColor, borderRadius: borderRadius),
                 height: double.infinity,
                 width: width ?? MediaQuery.of(context).size.width / 1.4,
                 child: body),