+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.4
+
+* Add sheet border radius parameter.
+* Add sheet color parameter.
+* Updated dependencies.
+* Fixed minor syntax errors.
+
 ## 1.0.3
 * Add doc.